% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaSimulation.R
\name{gamma.sim}
\alias{gamma.sim}
\title{Gamma Simulation Function}
\usage{
\method{gamma}{sim}(n, p, alpha, beta, ...)
}
\arguments{
\item{n}{sample size}

\item{p}{vector of percentiles of interest}

\item{alpha}{shape parameter for gamma function}

\item{beta}{scale parameter for gamma function}

\item{...}{Only included to prevent issues with R package}
}
\value{
absolute and relative bias of areas, means, variances, MSE, RSE
}
\description{
Gamma Simulation Function
}
\examples{
\dontrun{gamma.sim(3,c(.5,.6,.7,.8,.9,.95,.975),2,5)}
}
