% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalSimulation.R
\name{normal.sim}
\alias{normal.sim}
\title{Normal Simulation Function}
\usage{
normal.sim(n, p, stdev, mean)
}
\arguments{
\item{n}{sample size}

\item{p}{vector of percentiles of interest}

\item{stdev}{standard deviation of normal distribution}

\item{mean}{mean of normal distribution}
}
\value{
absolute and relative bias of areas, means, variances, MSE, RSE
}
\description{
Normal Simulation Function
}
\examples{
\dontrun{normal.sim(3,c(.5,.6,.7,.8,.9,.95,.975),1,1)}
}
